ingest/requirements.txt
Telethon==1.42.0
python-dotenv==1.2.1
pymongo==4.16.0
